pyproject.toml
[project]
name = "teamtask-server"
version = "0.1.0"
requires-python = ">=3.12.10"
dependencies = [
    "aiosqlite>=0.22.0",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.126.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
